Worked on Design part little left

+++ src/design/Employee.java
@@ -13,14 +13,15 @@ public interface Employee {
 	
 	//employeeName() will return employee name
 	public String employeeName();
-	
+
 	//assignDepartment() will assign employee to departments
     public void assignDepartment();
 	
 	//calculate employee salary
-	public int calculateSalary();
+	public double calculateSalary();
 	
 	//employee benefit
 	public void benefitLayout();
+
 
 }
